backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
google-generativeai==0.8.3
python-dotenv==1.0.1
PyMuPDF==1.25.3
pydantic==2.9.2
pydantic-settings==2.6.0
faiss-cpu==1.7.4
numpy==1.24.3

